refactor: split 3 big files into focused modules + extract cross-cutting helpers
ships.ts → turretFactory.ts (turret ECS creation extracted)
damage.ts → damageMath.ts (pure calc) + damageApplication.ts (mutation/callbacks)
systems.ts → pipeline.ts (declarative step ordering, measurement runner)
New helpers: tempVectors.ts (scratch Vector3/Quat pool), physicsBodyManager.ts (body lifecycle)

All imports backward-compatible; barrel re-exports preserve the original public API.
909 tests pass, typecheck clean.
